+++ browser-extension/src/entrypoints/posts/Report/Exports/reportDocx.ts
@@ -12,7 +12,7 @@ import {
   WidthType,
 } from "docx";
 import { Post } from "@/shared/model/post/Post";
-import { ReportQueryData } from "./BuildReport";
+import { ReportQueryData } from "../Stepper/BuildReport";
 import { getSocialNetworkName } from "@/shared/utils/post-util";
 import { SocialNetworkName } from "@/shared/model/SocialNetworkName";
 import {
