+++ browser-extension/src/entrypoints/posts/Report/Stepper/Step1Plateforme.tsx
