+++ browser-extension/src/entrypoints/posts/Report/Stepper/StepHeader.tsx
